I removed the default properties on entity and dto, I removed all implementations in my code, I renamed my entity class and I removed commented codes.
