+++ src/DwitTech.NotificationService.WebApi/Controllers/EmailController.cs
@@ -4,23 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace DwitTech.NotificationService.WebApi.Controllers
 {
-    [Route("api/[controller]")]
-    [ApiController]
-    public class EmailController : ControllerBase
+    public class EmailController : BaseController
     {
-        private readonly IEmailService _emailService;
-
-        public EmailController(IEmailService emailService)
-        {
-            _emailService = emailService;
-        }
-
-        [HttpPost]
-        public IActionResult SendEmail(EmailDto request)
-        {
-            _emailService.SendEmail(request);
-
-            return Ok();
-        }
+        
     }
 }
